+++ src/lib/lambda-config.ts
@@ -1,0 +1,7 @@
+/**
+ * Created by EGomez on 5/19/17.
+ */
+
+export interface LambdaConfig {
+    allowNullInjection:boolean;
+}