+++ src/lib/lambda-model.ts
@@ -2,9 +2,19 @@
  * Created by EGomez on 2/23/17.
  */
 
+import PathModel from "./path-model";
+import PathParamModel from "./path-param-model";
+import HttpVerbModel from "./http-verb-model";
+import {LambdaConfig} from "./lambda-config";
+
 export default class LambdaModel {
     private _name;
     private _instance;
+    private _paths:Array<PathModel> = [];
+    private _basePath:string;
+    private _pathParams:PathParamModel[] = [];
+    private _httpVerbs:HttpVerbModel[] = [];
+    private _config:LambdaConfig;
 
     /**
      * API Properties
@@ -33,13 +43,6 @@ export default class LambdaModel {
     private _bodyProperty;
     private _methodProperty;
 
-    private _getMethod;
-    private _postMethod;
-    private _patchMethod;
-    private _headMethod;
-    private _optionsMethod;
-    private _putMethod;
-    private _deleteMethod;
     private _anyMethod;
 
     get name() {
@@ -115,62 +118,6 @@ export default class LambdaModel {
         this._contextProperty = value;
     }
 
-    get getMethod() {
-        return this._getMethod;
-    }
-
-    set getMethod(value) {
-        this._getMethod = value;
-    }
-
-    get postMethod() {
-        return this._postMethod;
-    }
-
-    set postMethod(value) {
-        this._postMethod = value;
-    }
-
-    get patchMethod() {
-        return this._patchMethod;
-    }
-
-    set patchMethod(value) {
-        this._patchMethod = value;
-    }
-
-    get headMethod() {
-        return this._headMethod;
-    }
-
-    set headMethod(value) {
-        this._headMethod = value;
-    }
-
-    get optionsMethod() {
-        return this._optionsMethod;
-    }
-
-    set optionsMethod(value) {
-        this._optionsMethod = value;
-    }
-
-    get putMethod() {
-        return this._putMethod;
-    }
-
-    set putMethod(value) {
-        this._putMethod = value;
-    }
-
-    get deleteMethod() {
-        return this._deleteMethod;
-    }
-
-    set deleteMethod(value) {
-        this._deleteMethod = value;
-    }
-
     get anyMethod() {
         return this._anyMethod;
     }
@@ -225,6 +172,46 @@ export default class LambdaModel {
 
     set preLambdaCallbackMethod(value: string) {
         this._preLambdaCallbackMethod = value;
+    }
+
+    get paths(): Array<PathModel> {
+        return this._paths;
+    }
+
+    set paths(value: Array<PathModel>) {
+        this._paths = value;
+    }
+
+    get basePath(): string {
+        return this._basePath;
+    }
+
+    set basePath(value: string) {
+        this._basePath = value;
+    }
+
+    get pathParams(): PathParamModel[] {
+        return this._pathParams;
+    }
+
+    set pathParams(value: PathParamModel[]) {
+        this._pathParams = value;
+    }
+
+    get httpVerbs(): HttpVerbModel[] {
+        return this._httpVerbs;
+    }
+
+    set httpVerbs(value: HttpVerbModel[]) {
+        this._httpVerbs = value;
+    }
+
+    get config(): LambdaConfig {
+        return this._config;
+    }
+
+    set config(value: LambdaConfig) {
+        this._config = value;
     }
 
 }